Add files via upload
Virtual AI painter using OpenCV and Mediapipe is an application that tracks the movement of an object. Using this tracking feature the user can draw on the screen by moving the object (which in our project is the human hand) in the air, in front of the webcam. This real time webcam data generated by tracking the movement of the object helps the user to draw simple things which are both interesting and challenging.